+++ app/src/main/java/com/example/jamminbaru/ControllerCreator/ProfileCreator.java
@@ -1,11 +1,11 @@
 package com.example.jamminbaru.ControllerCreator;
 
 import static com.example.jamminbaru.creator.LoginCreator.creator;
-import static com.example.jamminbaru.user.LoginUser.user;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.content.ContextCompat;
 
 import android.content.ContentResolver;
 import android.content.Intent;
@@ -21,29 +21,33 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.jamminbaru.ControllerUser.ProfileUser;
+import com.example.jamminbaru.Model.ModelProfileCreator;
 import com.example.jamminbaru.Model.ModelProfileUser;
 import com.example.jamminbaru.R;
 import com.example.jamminbaru.creator.LoginCreator;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.OnProgressListener;
 import com.google.firebase.storage.StorageReference;
 import com.google.firebase.storage.UploadTask;
+import com.squareup.picasso.Picasso;
 
 public class ProfileCreator extends AppCompatActivity {
     ImageButton home;
     TextView usernameTag;
-    Button logoutDirectLogin;
+    Button logoutDirectLogin, upload_btn, delete_btn;
 
-    private Button upload_btn;
     private ImageView imageView;
     private ProgressBar progressBar;
     final private String displayUsername = creator.getPhoneTxt();
-    private DatabaseReference root = FirebaseDatabase.getInstance().getReference().child("Creator").child(displayUsername).child("Profile_Image");
-    private StorageReference reference = FirebaseStorage.getInstance().getReference().child("Creator").child(displayUsername).child("Profile_Image");
+    private final DatabaseReference root = FirebaseDatabase.getInstance().getReference().child("Creator").child(displayUsername).child("Profile_Image");
+    private final StorageReference reference = FirebaseStorage.getInstance().getReference().child("Creator").child(displayUsername).child("Profile_Image");
     private Uri imageUri;
 
     @Override
@@ -55,8 +59,9 @@ public class ProfileCreator extends AppCompatActivity {
         usernameTag = findViewById(R.id.usernameTag);
         usernameTag.setText(displayUsername);
 
-        logoutDirectLogin = (Button) findViewById(R.id.logoutDirectLogin);
+        logoutDirectLogin = findViewById(R.id.logoutDirectLogin);
         upload_btn = findViewById(R.id.upload_btn);
+        delete_btn = findViewById(R.id.delete_btn);
         imageView = findViewById(R.id.profileimg);
         progressBar = findViewById(R.id.progressBar);
         progressBar.setVisibility(View.INVISIBLE);
@@ -99,6 +104,59 @@ public class ProfileCreator extends AppCompatActivity {
                 startActivity(intent);
             }
         });
+
+        delete_btn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                root.addListenerForSingleValueEvent(new ValueEventListener() {
+                    @Override
+                    public void onDataChange(@NonNull DataSnapshot snapshot) {
+                        if (snapshot.exists()) {
+                            // data Link URL image ditemukan, hapus data tersebut
+                            root.removeValue().addOnSuccessListener(new OnSuccessListener<Void>() {
+                                @Override
+                                public void onSuccess(Void unused) {
+                                    // data Link URL image berhasil dihapus, ubah gambar di ImageView
+                                    imageView.setImageDrawable(ContextCompat.getDrawable(ProfileCreator.this, R.drawable.ic_baseline_add_photo_alternate_24));
+                                    Toast.makeText(ProfileCreator.this, "Delete Picture Successfully", Toast.LENGTH_SHORT).show();
+                                }
+                            });
+                        } else {
+                            // data Link URL image tidak ditemukan, munculkan pesan error
+                            Toast.makeText(ProfileCreator.this, "No Image Found", Toast.LENGTH_SHORT).show();
+                        }
+                    }
+
+                    @Override
+                    public void onCancelled(@NonNull DatabaseError error) {
+                        // error handling
+                    }
+                });
+            }
+        });
+
+        getUserInfo();
+    }
+
+    private void getUserInfo(){
+        root.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                if (snapshot.exists() && snapshot.getChildrenCount() > 0){
+                    if (snapshot.hasChild("imageUrl")){
+                        // link image foto ditemukan, tampilkan gambar
+                        String image = snapshot.child("imageUrl").getValue().toString();
+                        Picasso.get().load(image).into(imageView);
+                    }
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+
+            }
+        });
+
     }
 
     @Override
@@ -119,10 +177,9 @@ public class ProfileCreator extends AppCompatActivity {
                 fileRef.getDownloadUrl().addOnSuccessListener(new OnSuccessListener<Uri>() {
                     @Override
                     public void onSuccess(Uri uri) {
-
-                        ModelProfileUser model = new ModelProfileUser(uri.toString());
-                        String modelId = root.push().getKey();
-                        root.child(modelId).setValue(model);
+                        String downloadUrl = uri.toString();
+                        ModelProfileCreator model = new ModelProfileCreator(downloadUrl);
+                        root.setValue(model);
                         progressBar.setVisibility(View.INVISIBLE);
                         Toast.makeText(ProfileCreator.this, "Uploaded Successfully!", Toast.LENGTH_SHORT).show();
                     }
