--- app/src/main/java/com/example/jamminbaru/DAOUser.java
@@ -1,18 +1,0 @@
-package com.example.jamminbaru;
-
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-
-public class DAOUser {
-//    private DatabaseReference databaseReference;
-//
-//    public DAOUser(){
-//        FirebaseDatabase db = FirebaseDatabase.getInstance();
-//        databaseReference = db.getReference(User.class.getSimpleName());
-//    }
-//
-//    public Task<Void> add(User usr){
-//        return databaseReference.push().setValue(usr);
-//    }
-}
