+++ app/src/main/java/com/example/jamminbaru/Model/ModelProfileUser.java
@@ -3,10 +3,6 @@ package com.example.jamminbaru.Model;
 public class ModelProfileUser {
     private String imageUrl;
 
-//    public ModelProfileUser(){
-//
-//    }
-
     public ModelProfileUser(String imageUrl){
         this.imageUrl = imageUrl;
     }
