--- app/src/main/java/com/example/jamminbaru/User.java
@@ -1,21 +1,0 @@
-package com.example.jamminbaru;
-
-public class User {
-    private String phoneTxt = "nama";
-
-    public User(String phoneTxt) {
-        this.phoneTxt = phoneTxt;
-    }
-
-    public String getPhoneTxt() {
-        return phoneTxt;
-    }
-
-    public void setPhoneTxt(String phoneTxt) {
-        this.phoneTxt = phoneTxt;
-    }
-
-//    public void addUser(String phoneTxt){
-//        Array[0] = phoneTxt;
-//    }
-}
