+++ app/src/main/res/layout/activity_home2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home2">
+    tools:context=".ControllerUser.HomeUser">
 
     <androidx.appcompat.widget.SearchView
         android:id="@+id/searchview"
