+++ app/src/main/res/layout/activity_profile_creator.xml
@@ -8,30 +8,40 @@
 
     <ImageView
         android:id="@+id/profileimg"
-        android:layout_width="118dp"
-        android:layout_height="91dp"
+        android:layout_width="244dp"
+        android:layout_height="225dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.235"
+        app:layout_constraintVertical_bias="0.063"
         app:srcCompat="@drawable/ic_baseline_add_photo_alternate_24" />
 
     <Button
         android:id="@+id/upload_btn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="162dp"
-        android:layout_marginTop="61dp"
-        android:layout_marginEnd="156dp"
-        android:layout_marginBottom="35dp"
-        android:text="UPLOAD"
-        app:layout_constraintBottom_toTopOf="@+id/progressBar"
+        android:layout_marginStart="159dp"
+        android:layout_marginTop="32dp"
+        android:layout_marginEnd="159dp"
+        android:text="UPLOAD IMAGE"
+        app:layout_constraintBottom_toTopOf="@+id/delete_btn"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/profileimg" />
+
+    <Button
+        android:id="@+id/delete_btn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="133dp"
+        android:layout_marginEnd="134dp"
+        android:layout_marginBottom="17dp"
+        android:text="DELETE IMAGE"
+        app:layout_constraintBottom_toTopOf="@+id/progressBar"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/upload_btn" />
 
     <ProgressBar
         android:id="@+id/progressBar"
@@ -39,7 +49,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="181dp"
-        android:layout_marginTop="34dp"
         android:layout_marginEnd="182dp"
         android:layout_marginBottom="33dp"
         android:indeterminateTint="@color/purple_200"
@@ -47,7 +56,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/upload_btn" />
+        app:layout_constraintTop_toBottomOf="@+id/delete_btn" />
 
     <TextView
         android:id="@+id/textView"
