+++ app/src/main/res/layout/activity_profile_user.xml
@@ -6,18 +6,6 @@
     android:layout_height="match_parent"
     tools:context=".ControllerUser.ProfileUser">
 
-    <ImageView
-        android:id="@+id/profileimg"
-        android:layout_width="118dp"
-        android:layout_height="91dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.235"
-        app:srcCompat="@drawable/ic_baseline_add_photo_alternate_24" />
-
 
     <!--    <TextView-->
     <!--        android:id="@+id/usernameTag"-->
@@ -32,6 +20,17 @@
     <!--        app:layout_constraintStart_toEndOf="@+id/textView"-->
     <!--        app:layout_constraintTop_toTopOf="parent"-->
     <!--        app:layout_constraintVertical_bias="0.422" />-->
+
+    <ImageView
+        android:id="@+id/profileimg"
+        android:layout_width="244dp"
+        android:layout_height="225dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.063"
+        app:srcCompat="@drawable/ic_baseline_add_photo_alternate_24" />
 
     <Button
         android:id="@+id/logoutDirectLogin"
@@ -85,16 +84,27 @@
         android:id="@+id/upload_btn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="162dp"
-        android:layout_marginTop="61dp"
-        android:layout_marginEnd="156dp"
-        android:layout_marginBottom="35dp"
-        android:text="UPLOAD"
-        app:layout_constraintBottom_toTopOf="@+id/progressBar"
+        android:layout_marginStart="159dp"
+        android:layout_marginTop="32dp"
+        android:layout_marginEnd="159dp"
+        android:text="UPLOAD IMAGE"
+        app:layout_constraintBottom_toTopOf="@+id/delete_btn"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/profileimg" />
+
+    <Button
+        android:id="@+id/delete_btn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="133dp"
+        android:layout_marginEnd="134dp"
+        android:layout_marginBottom="17dp"
+        android:text="DELETE IMAGE"
+        app:layout_constraintBottom_toTopOf="@+id/progressBar"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/upload_btn" />
 
     <ProgressBar
         android:id="@+id/progressBar"
@@ -102,7 +112,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="181dp"
-        android:layout_marginTop="34dp"
         android:layout_marginEnd="182dp"
         android:layout_marginBottom="33dp"
         android:indeterminateTint="@color/purple_200"
@@ -110,7 +119,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/upload_btn" />
+        app:layout_constraintTop_toBottomOf="@+id/delete_btn" />
 
     <TextView
         android:id="@+id/textView"
@@ -135,5 +144,15 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.713" />
+
+<!--    <TextView-->
+<!--        android:id="@+id/profileChangeBtn"-->
+<!--        android:layout_width="wrap_content"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        android:layout_marginTop="40dp"-->
+<!--        android:text="Change Profile Picture"-->
+<!--        app:layout_constraintEnd_toEndOf="parent"-->
+<!--        app:layout_constraintStart_toStartOf="parent"-->
+<!--        app:layout_constraintTop_toBottomOf="@+id/profileimg" />-->
 
 </androidx.constraintlayout.widget.ConstraintLayout>
